added tps display to server and some other changes i can't remember
Former-commit-id: ae82376266db5169075b788e3e6bc49ca031404d [formerly 797fb10be67d9f072b3dd485c6314aa23559e12e]
Former-commit-id: ec8794b71932715d2360f1a254eea3dda448f701

+++ bw4t/bw4t-client/src/main/java/nl/tudelft/bw4t/client/gui/menu/HallwayMenu.java
@@ -34,8 +34,8 @@ public class HallwayMenu {
 		BasicMenuOperations.addSectionTitleToPopupMenu("Command my robot to: ", gui.getjPopupMenu());
 
 		JMenuItem menuItem = new JMenuItem("Go to here");
-		menuItem.addActionListener(new GotoPositionActionListener(new Point(gui.getSelectedLocation().x
-				/ set.getScale(), gui.getSelectedLocation().y / set.getScale()), gui.getController()));
+		menuItem.addActionListener(new GotoPositionActionListener(new Point((int) (gui.getSelectedLocation().x / set
+				.getScale()), (int) (gui.getSelectedLocation().y / set.getScale())), gui.getController()));
 		gui.getjPopupMenu().add(menuItem);
 
 		if (holdingID != null) {
@@ -68,7 +68,8 @@ public class HallwayMenu {
 			for (Zone roomInfo : cmc.getRooms()) {
 				menuItem = new JMenuItem(roomInfo.getName());
 				menuItem.addActionListener(new MessageSenderActionListener(new BW4TMessage(MessageType.hasColor,
-						roomInfo.getName(), ColorTranslator.translate2ColorString(entityColor), null), gui.getController()));
+						roomInfo.getName(), ColorTranslator.translate2ColorString(entityColor), null), gui
+						.getController()));
 				submenu.add(menuItem);
 			}
 		}
